Refactored CameraPoll location
