+++ src/org/usfirst/frc4904/robot/commands/camera/CameraPoll.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc4904.robot.commands.shooter;
+package org.usfirst.frc4904.robot.commands.camera;
 
 
 import java.io.BufferedReader;
